Falls Netwerk nicht vorhanden erstellt sich die Navigation anhand SharedPreferences neu

+++ app/src/main/java/de/haertel/hawapp/campusnoticeboard/impl/navigationMenu/data/NavigationMenuDataHandler.java
@@ -4,6 +4,10 @@ import android.app.Activity;
 import android.arch.lifecycle.LifecycleOwner;
 import android.arch.lifecycle.Observer;
 import android.arch.lifecycle.ViewModelProviders;
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.FragmentActivity;
@@ -15,6 +19,8 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -29,6 +35,9 @@ import de.haertel.hawapp.campusnoticeboard.impl.navigationMenu.presentation.Expa
 import de.haertel.hawapp.campusnoticeboard.impl.navigationMenu.presentation.ExpandedMenuModel;
 
 public class NavigationMenuDataHandler {
+    private final String MENU_ENTRY_STORE_KEY = "menuEntryPreference";
+    private final String MY_PREFERENCES = "MyPreferences";
+
     ExpandableListAdapter mFacultyMenuAdapter;
     ExpandableListAdapter mGeneralMenuAdapter;
     ExpandableListView facultyMenuExpandableList;
@@ -56,22 +65,34 @@ public class NavigationMenuDataHandler {
         //navigationView.setNavigationItemSelectedListener(this);
         menuEntryViewModel = ViewModelProviders.of((FragmentActivity) mainActivity).get(MenuEntryViewModel.class);
 
+        if(! _isNetworkAvailable()){
+            List<MenuEntry> storedMenuEntries =_fetchMenuEntries();
+            if (storedMenuEntries != null) {
+                _prepareNavigationMenu(mainActivity.getString(R.string.facultyTopic), storedMenuEntries);
+                _prepareNavigationMenu(mainActivity.getString(R.string.generalTopic), storedMenuEntries);
+            }
+        }
+
         mDatabase.addValueEventListener(new ValueEventListener() {
             @Override
             @SuppressWarnings("unchecked")
             public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                arrayList = (ArrayList<HashMap<String, String>>) dataSnapshot.getValue();
-                menuEntryViewModel.deleteAllMenuEntries();
-                _populateDatabase();
-                menuEntryViewModel.getAllMenuEntries().observe((LifecycleOwner) mainActivity, new Observer<List<MenuEntry>>() {
-                    @Override
-                    public void onChanged(@Nullable final List<MenuEntry> menuEntries) {
-                        if (menuEntries != null){
-                            _prepareNavigationMenu(mainActivity.getString(R.string.facultyTopic), menuEntries);
-                            _prepareNavigationMenu(mainActivity.getString(R.string.generalTopic), menuEntries);
+                if (_isNetworkAvailable()) {
+                    arrayList = (ArrayList<HashMap<String, String>>) dataSnapshot.getValue();
+                    menuEntryViewModel.deleteAllMenuEntries();
+                    _populateDatabase();
+                    menuEntryViewModel.getAllMenuEntries().observe((LifecycleOwner) mainActivity, new Observer<List<MenuEntry>>() {
+                        @Override
+                        public void onChanged(@Nullable final List<MenuEntry> menuEntries) {
+                            if (menuEntries != null && _isNetworkAvailable()) {
+                                _storeMenuEntries(menuEntries);
+                                _prepareNavigationMenu(mainActivity.getString(R.string.facultyTopic), menuEntries);
+                                _prepareNavigationMenu(mainActivity.getString(R.string.generalTopic), menuEntries);
+                            }
                         }
-                    }
-                });
+                    });
+                }
+
             }
 
             @Override
@@ -79,6 +100,36 @@ public class NavigationMenuDataHandler {
 
             }
         });
+    }
+
+    private boolean _storeMenuEntries(List<MenuEntry> pMenuEntries)    {
+        boolean returnValue = false;
+        SharedPreferences sharedPreferences;
+        SharedPreferences.Editor editor;
+        sharedPreferences = mainActivity.getSharedPreferences(MY_PREFERENCES, Context.MODE_PRIVATE);
+
+        Gson gson = new Gson();
+        String json = gson.toJson(pMenuEntries);
+
+        editor = sharedPreferences.edit();
+        editor.remove(MENU_ENTRY_STORE_KEY).commit();
+        editor.putString(MENU_ENTRY_STORE_KEY, json);
+        returnValue = editor.commit();
+
+        return returnValue;
+    }
+    private List<MenuEntry> _fetchMenuEntries(){
+        SharedPreferences sharedPreferences;
+        sharedPreferences = mainActivity.getSharedPreferences(MY_PREFERENCES, Context.MODE_PRIVATE);
+        if (sharedPreferences != null){
+            Gson gson = new Gson();
+            String response= sharedPreferences.getString(MENU_ENTRY_STORE_KEY , "");
+            ArrayList<MenuEntry> menuEntryList = gson.fromJson(response,
+                    new TypeToken<List<MenuEntry>>(){}.getType());
+            return menuEntryList;
+        }else {
+            return null;
+        }
     }
 
     private void _populateDatabase() {
@@ -177,7 +228,6 @@ public class NavigationMenuDataHandler {
                 return model1.getMenuName().toLowerCase().compareTo(model2.getMenuName().toLowerCase());
             }
         });
-        System.out.println();
         for (List<String> list : navigationMenuChildList.values()) {
             Collections.sort(list, new Comparator<String>() {
                 @Override
@@ -186,7 +236,6 @@ public class NavigationMenuDataHandler {
                 }
             });
         }
-        System.out.println();
 
     }
 
@@ -225,4 +274,10 @@ public class NavigationMenuDataHandler {
         });
     }
 
+    private boolean _isNetworkAvailable() {
+        ConnectivityManager connectivityManager
+                = (ConnectivityManager) mainActivity.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
+        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
+    }
 }
